Rename precice driver files
