+++ include/exadg/structure/spatial_discretization/operator.h
@@ -301,6 +301,7 @@ public:
   std::shared_ptr<ContainerInterfaceData<dim, dim, Number>>
   get_container_interface_data_dirichlet();
 
+  // TODO: we currently need this function public for precice-based FSI
   unsigned int
   get_dof_index() const;
